+++ include/engine/routing_algorithms.hpp
@@ -30,11 +30,12 @@ class RoutingAlgorithmsInterface
     virtual InternalRouteResult
     DirectShortestPathSearch(const PhantomNodes &phantom_node_pair) const = 0;
 
-    virtual std::pair<std::vector<EdgeDuration>, std::vector<EdgeDistance>>
-    ManyToManySearch(const std::vector<PhantomNode> &phantom_nodes,
-                     const std::vector<std::size_t> &source_indices,
-                     const std::vector<std::size_t> &target_indices,
-                     const bool calculate_distance) const = 0;
+    virtual std::
+        tuple<std::vector<EdgeWeight>, std::vector<EdgeDuration>, std::vector<EdgeDistance>>
+        ManyToManySearch(const std::vector<PhantomNode> &phantom_nodes,
+                         const std::vector<std::size_t> &source_indices,
+                         const std::vector<std::size_t> &target_indices,
+                         const bool calculate_distance) const = 0;
 
     virtual routing_algorithms::SubMatchingList
     MapMatching(const routing_algorithms::CandidateLists &candidates_list,
@@ -83,11 +84,12 @@ template <typename Algorithm> class RoutingAlgorithms final : public RoutingAlgo
     InternalRouteResult
     DirectShortestPathSearch(const PhantomNodes &phantom_nodes) const final override;
 
-    virtual std::pair<std::vector<EdgeDuration>, std::vector<EdgeDistance>>
-    ManyToManySearch(const std::vector<PhantomNode> &phantom_nodes,
-                     const std::vector<std::size_t> &source_indices,
-                     const std::vector<std::size_t> &target_indices,
-                     const bool calculate_distance) const final override;
+    virtual std::
+        tuple<std::vector<EdgeWeight>, std::vector<EdgeDuration>, std::vector<EdgeDistance>>
+        ManyToManySearch(const std::vector<PhantomNode> &phantom_nodes,
+                         const std::vector<std::size_t> &source_indices,
+                         const std::vector<std::size_t> &target_indices,
+                         const bool calculate_distance) const final override;
 
     routing_algorithms::SubMatchingList
     MapMatching(const routing_algorithms::CandidateLists &candidates_list,
@@ -192,7 +194,7 @@ inline routing_algorithms::SubMatchingList RoutingAlgorithms<Algorithm>::MapMatc
 }
 
 template <typename Algorithm>
-std::pair<std::vector<EdgeDuration>, std::vector<EdgeDistance>>
+std::tuple<std::vector<EdgeWeight>, std::vector<EdgeDuration>, std::vector<EdgeDistance>>
 RoutingAlgorithms<Algorithm>::ManyToManySearch(const std::vector<PhantomNode> &phantom_nodes,
                                                const std::vector<std::size_t> &_source_indices,
                                                const std::vector<std::size_t> &_target_indices,